Replaced scenario with new implementation
