+++ src/main/scala/com/canoe/telegram/scenarios/Scenario.scala
@@ -1,213 +1,167 @@
 package com.canoe.telegram.scenarios
 
-import cats.{Applicative, MonadError}
-import com.canoe.telegram.models.messages.TelegramMessage
+import cats.{Applicative, Monad}
 import fs2.{Pipe, Pull, Stream}
+
+import Scenario._
+
+sealed trait Scenario[F[_], -I, +O] extends Pipe[F, I, O] {
+  def apply(input: Stream[F, I]): Stream[F, O] =
+    loop(this, input, Nil).collect {
+      case (Matched(o), _) => o
+    }
+
+  def flatMap[I2 <: I, O2](fn: O => Scenario[F, I2, O2]): Scenario[F, I2, O2] =
+    Bind(this, fn)
+
+  def map[O2](fn: O => O2): Scenario[F, I, O2] = Mapped(this, fn)
+
+  def cancelOn[I2 <: I](p: I2 => Boolean): Scenario[F, I2, O] = Cancellable(this, p)
+
+  def tolerateN[I2 <: I, A](n: Int)(fn: I2 => F[A]): Scenario[F, I2, O] =
+    Tolerate(this, Some(n), fn)
+
+  def tolerate[I2 <: I, A](fn: I2 => F[A]): Scenario[F, I2, O] = tolerateN(1)(fn)
+
+  def tolerateAll[I2 <: I, A](fn: I2 => F[A]): Scenario[F, I2, O] =
+    Tolerate(this, None, fn)
+}
+
+final case class Eval[F[_], I, A](fa: F[A]) extends Scenario[F, I, A]
+
+final case class Next[F[_], A](p: A => Boolean) extends Scenario[F, A, A]
+
+final case class Start[F[_], A](p: A => Boolean) extends Scenario[F, A, A]
+
+final case class Bind[F[_], I, O1, O2](scenario: Scenario[F, I, O1],
+                                       fn: O1 => Scenario[F, I, O2])
+    extends Scenario[F, I, O2]
+
+final case class Mapped[F[_], I, O1, O2](scenario: Scenario[F, I, O1],
+                                         fn: O1 => O2)
+    extends Scenario[F, I, O2]
+
+final case class Cancellable[F[_], I, O](scenario: Scenario[F, I, O],
+                                         cancelOn: I => Boolean)
+    extends Scenario[F, I, O]
+
+final case class Tolerate[F[_], I, O, A](scenario: Scenario[F, I, O],
+                                         limit: Option[Int],
+                                         fn: I => F[A])
+    extends Scenario[F, I, O]
 
 object Scenario {
 
-  def receive[F[_]](p: PartialFunction[TelegramMessage, Boolean]): Scenario[F, TelegramMessage] =
-    Receive(p.applyOrElse(_, (_: Any) => false))
+  def eval[F[_], I, A](fa: F[A]): Scenario[F, I, A] = Eval(fa)
 
-  def expect[F[_]](p: PartialFunction[TelegramMessage, Boolean]): Scenario[F, TelegramMessage] =
-    Expect(p.applyOrElse(_, (_: Any) => false))
+  def start[F[_], I](p: I => Boolean): Scenario[F, I, I] = Start(p)
 
-  def eval[F[_], A](fa: F[A]): Scenario[F, A] =
-    Action(fa)
+  def next[F[_], I](p: I => Boolean): Scenario[F, I, I] = Next(p)
 
-  def unit[F[_]](implicit F: Applicative[F]) =
-    Action(F.unit)
-
-  def drain[F[_], A]: Scenario[F, A] =
-    new Scenario[F, A] {
-      def fulfill(messages: Stream[F, TelegramMessage]): Stream[F, (Result[A], Stream[F, TelegramMessage])] =
-        Stream.empty
-    }
-
-  implicit def monadErrorInstance[F[_]](implicit F: Applicative[F]): MonadError[Scenario[F, ?], TelegramMessage] =
-    new MonadError[Scenario[F, ?], TelegramMessage] {
-      def raiseError[A](e: TelegramMessage): Scenario[F, A] =
-        new Scenario[F, A] {
-          def fulfill(messages: Stream[F, TelegramMessage]): Stream[F, (Result[A], Stream[F, TelegramMessage])] =
-            Stream(Missed(e) -> messages)
-        }
-
-      def handleErrorWith[A](fa: Scenario[F, A])(f: TelegramMessage => Scenario[F, A]): Scenario[F, A] = {
-        new Scenario[F, A] {
-          def fulfill(messages: Stream[F, TelegramMessage]): Stream[F, (Result[A], Stream[F, TelegramMessage])] =
-            fa.fulfill(messages).flatMap {
-              case (Missed(m), rest) =>
-                f(m).fulfill(rest)
-
-              case success => Stream(success)
-            }
-        }
-      }
-
-      override def pure[A](x: A): Scenario[F, A] =
-        Action(F.pure(x))
-
-      override def flatMap[A, B](fa: Scenario[F, A])(f: A => Scenario[F, B]): Scenario[F, B] =
-        fa.flatMap(f)
-
-      override def tailRecM[A, B](a: A)(f: A => Scenario[F, Either[A, B]]): Scenario[F, B] =
-        f(a).flatMap {
-          case Left(a) => tailRecM(a)(f)
-          case Right(b) => Action(F.pure(b))
-        }
-    }
-
-  sealed trait Result[+A] {
-    def map[B](f: A => B): Result[B] = this match {
-      case Matched(a) => Matched(f(a))
-      case same @ Missed(_) => same
+  sealed trait Result[+E, +A] {
+    def map[B](f: A => B): Result[E, B] = this match {
+      case Matched(a)          => Matched(f(a))
+      case same @ Missed(_)    => same
       case same @ Cancelled(_) => same
     }
   }
 
-  case class Matched[A](a: A) extends Result[A]
-  case class Missed(message: TelegramMessage) extends Result[Nothing]
-  case class Cancelled(message: TelegramMessage) extends Result[Nothing]
+  final case class Matched[A](a: A) extends Result[Nothing, A]
+  final case class Missed[E](message: E) extends Result[E, Nothing]
+  final case class Cancelled[E](message: E) extends Result[E, Nothing]
 
-}
+  private def loop[F[_], I, O](
+    scenario: Scenario[F, I, O],
+    input: Stream[F, I],
+    cancelTokens: List[I => Boolean]
+  ): Stream[F, (Result[I, O], Stream[F, I])] = {
+    scenario match {
+      case Eval(fa) =>
+        Stream.eval(fa).map(o => Matched(o) -> input)
 
-import Scenario._
+      case Start(p: (I => Boolean)) =>
+        def go(in: Stream[F, I]): Pull[F, (Result[I, O], Stream[F, I]), Unit] =
+          in.dropWhile(!p(_)).pull.uncons1.flatMap {
+            case Some((m, rest)) =>
+              Pull.output1(Matched(m.asInstanceOf[O]) -> rest) >> go(rest)
 
-sealed abstract class Scenario[F[_], A] extends Pipe[F, TelegramMessage, A] {
-  self =>
+            case None => Pull.done
+          }
 
-  def apply(messages: Stream[F, TelegramMessage]): Stream[F, A] =
-    fulfill(messages).map(_._1).collect { case Matched(a) => a }
+        go(input).stream
 
-  def fulfill(messages: Stream[F, TelegramMessage]): Stream[F, (Result[A], Stream[F, TelegramMessage])]
+      case Next(p) =>
+        def go(in: Stream[F, I]): Pull[F, (Result[I, O], Stream[F, I]), Unit] =
+          in.pull.uncons1.flatMap {
+            case Some((m, rest)) =>
+              if (cancelTokens.exists(_(m)))
+                Pull.output1(Cancelled(m) -> rest)
+              else if (p(m)) Pull.output1(Matched(m.asInstanceOf[O]) -> rest)
+              else Pull.output1(Missed(m) -> rest)
 
-  def map[B](fn: A => B): Scenario[F, B] = new Scenario[F, B] {
-    def fulfill(messages: Stream[F, TelegramMessage]): Stream[F, (Result[B], Stream[F, TelegramMessage])] =
-      self.fulfill(messages).map {
-        case (a, rest) => (a.map(fn), rest)
-      }
-  }
+            case None => Pull.done
+          }
 
-  def flatMap[B](fn: A => Scenario[F, B]): Scenario[F, B] = {
-    new Scenario[F, B] {
-      def fulfill(messages: Stream[F, TelegramMessage]): Stream[F, (Result[B], Stream[F, TelegramMessage])] =
-        self.fulfill(messages).flatMap {
-          case (Matched(a), rest) => fn(a).fulfill(rest)
-          case (Missed(m), rest) => Stream(Missed(m) -> rest)
-          case (Cancelled(m), rest) => Stream(Cancelled(m) -> rest)
-        }
+        go(input).stream
 
-    }
-  }
+      case Cancellable(scenario, p) =>
+        loop(scenario, input, p :: cancelTokens)
 
-  def tolerate[B](fn: TelegramMessage => F[B]): Scenario[F, A] = tolerateN(1)(fn)
+      case Tolerate(scenario, limit, fn) =>
+        limit match {
+          case Some(n) if n <= 0 =>
+            loop(scenario, input, cancelTokens)
 
-  def tolerateN[B](n: Int)(fn: TelegramMessage => F[B]): Scenario[F, A] = {
-
-    def tolerateFulfill(n: Int, messages: Stream[F, TelegramMessage]): Stream[F, (Result[A], Stream[F, TelegramMessage])] =
-      if (n <= 0) self.fulfill(messages)
-      else self.fulfill(messages).flatMap {
-        case (Cancelled(_), rest) => self.fulfill(rest)
-        case (Missed(m), rest) => Stream.eval(fn(m)) >> tolerateFulfill(n - 1, rest)
-        case matched => Stream(matched)
-      }
-
-    new Scenario[F, A] {
-      def fulfill(messages: Stream[F, TelegramMessage]): Stream[F, (Result[A], Stream[F, TelegramMessage])] =
-        tolerateFulfill(n, messages)
-    }
-  }
-
-  // Doesn't keep the history of matched messages
-  // (second scenario starts matching from the point where the first one missed / was cancelled)
-  def or[B](other: => Scenario[F, B]): Scenario[F, Either[A, B]] =
-    new Scenario[F, Either[A, B]] {
-      def fulfill(messages: Stream[F, TelegramMessage]): Stream[F, (Result[Either[A, B]], Stream[F, TelegramMessage])] =
-        self.fulfill(messages).flatMap {
-          case (Matched(a), rest) =>
-            Stream(Matched(Left(a)) -> rest)
-
-          case (Missed(m), rest) =>
-            other.fulfill(rest.cons1(m)).map {
-              case (Matched(b), rest2) => Matched(Right(b)) -> rest2
-              case (Missed(m), rest2) => Missed(m) -> rest2
-              case (Cancelled(m), rest2) => Cancelled(m) -> rest2
-            }
-
-          // Try second scenario even if the first one was cancelled (should the cancellation message be read? [other.fulfill(rest.cons1(m)).map])
-          case (Cancelled(m), rest) =>
-            other.fulfill(rest.cons1(m)).map {
-              case (Matched(b), rest2) => Matched(Right(b)) -> rest2
-              case (Missed(m), rest2) => Missed(m) -> rest2
-              case (Cancelled(m), rest2) => Cancelled(m) -> rest2
+          case _ =>
+            loop(scenario, input, cancelTokens).flatMap {
+              case (Cancelled(m), rest) => Stream(Cancelled(m) -> rest)
+              case (Missed(m), rest) =>
+                Stream.eval(fn(m)) >> loop(
+                  Tolerate(scenario, limit.map(_ - 1), fn),
+                  rest,
+                  cancelTokens
+                )
+              case matched => Stream(matched)
             }
         }
+
+      case Bind(prev, fn) =>
+        loop(prev, input, cancelTokens).flatMap {
+          case (result, rest) =>
+            result match {
+              case Matched(a)         => loop(fn(a), rest, cancelTokens)
+              case Missed(message)    => Stream(Missed(message) -> rest)
+              case Cancelled(message) => Stream(Cancelled(message) -> rest)
+            }
+        }
+
+      case Mapped(prev, fn) =>
+        loop(prev, input, cancelTokens).map {
+          case (result, rest) => result.map(fn) -> rest
+        }
     }
-
-  def drain: Scenario[F, Unit] = flatMap(_ => Scenario.drain)
-}
-
-object Expect {
-
-  def any[F[_]]: Expect[F] = Expect(_ => true)
-
-  def collect[F[_]]: ExpectCollectPartiallyApplied[F] = new ExpectCollectPartiallyApplied[F]
-
-  final class ExpectCollectPartiallyApplied[F[_]](val dummy: Boolean = true) extends AnyVal {
-    def apply[A](p: PartialFunction[TelegramMessage, A]): Scenario[F, A] =
-      Expect(p.isDefinedAt).map(p)
   }
 
-}
+  implicit def monadInstance[F[_]: Applicative, I]: Monad[Scenario[F, I, ?]] =
+    new Monad[Scenario[F, I, ?]] {
+      def pure[A](x: A): Scenario[F, I, A] = Eval(Applicative[F].pure(x))
 
-final case class Expect[F[_]](p: TelegramMessage => Boolean) extends Scenario[F, TelegramMessage] {
+      def flatMap[A, B](
+        scenario: Scenario[F, I, A]
+      )(fn: A => Scenario[F, I, B]): Scenario[F, I, B] =
+        scenario.flatMap(fn)
 
-  def fulfill(messages: Stream[F, TelegramMessage]): Stream[F, (Result[TelegramMessage], Stream[F, TelegramMessage])] = {
+      override def map[A, B](
+        scenario: Scenario[F, I, A]
+      )(f: A => B): Scenario[F, I, B] =
+        scenario.map(f)
 
-    def go(input: Stream[F, TelegramMessage]): Pull[F, (Result[TelegramMessage], Stream[F, TelegramMessage]), Unit] =
-      input.pull.uncons1.flatMap {
-        case Some((m, rest)) =>
-          if (p(m)) Pull.output1(Matched(m) -> rest)
-          else Pull.output1(Missed(m) -> rest)
-
-        case None => Pull.done
-      }
-
-    go(messages).stream
-  }
-}
-
-object Receive {
-
-  def any[F[_]]: Receive[F] = Receive(_ => true)
-
-  def collect[F[_]]: ReceiveCollectPartiallyApplied[F] = new ReceiveCollectPartiallyApplied[F]
-
-  final class ReceiveCollectPartiallyApplied[F[_]](val dummy: Boolean = true) extends AnyVal {
-    def apply[A](p: PartialFunction[TelegramMessage, A]): Scenario[F, A] =
-      Receive(p.isDefinedAt).map(p)
-  }
-
-}
-
-final case class Receive[F[_]](p: TelegramMessage => Boolean) extends Scenario[F, TelegramMessage] {
-
-  def fulfill(messages: Stream[F, TelegramMessage]): Stream[F, (Result[TelegramMessage], Stream[F, TelegramMessage])] = {
-
-    def go(input: Stream[F, TelegramMessage]): Pull[F, (Result[TelegramMessage], Stream[F, TelegramMessage]), Unit] = {
-      input.dropWhile(!p(_)).pull.uncons1.flatMap {
-        case Some((m, rest)) =>
-          Pull.output1(Matched(m) -> rest) >> go(rest)
-
-        case None => Pull.done
-      }
+      def tailRecM[A, B](
+        a: A
+      )(f: A => Scenario[F, I, Either[A, B]]): Scenario[F, I, B] =
+        flatMap(f(a)) {
+          case Left(a)  => tailRecM(a)(f)
+          case Right(b) => Eval(Applicative[F].pure(b))
+        }
     }
-
-    go(messages).stream
-  }
-}
-
-final case class Action[F[_], A](fa: F[A]) extends Scenario[F, A] {
-  def fulfill(messages: Stream[F, TelegramMessage]): Stream[F, (Result[A], Stream[F, TelegramMessage])] =
-    Stream.eval(fa).map(Matched(_) -> messages)
 }
