--- src/test/scala/com/canoe/telegram/scenarios/ScenarioLawTests.scala
@@ -1,8 +1,0 @@
-package com.canoe.telegram.scenarios
-
-import org.scalatest.FunSuite
-import org.typelevel.discipline.scalatest.Discipline
-
-class ScenarioLawTests extends FunSuite with Discipline {
-
-}
