+++ src/test/scala/com/canoe/telegram/scenarios/ScenarioSpec.scala
@@ -3,7 +3,7 @@ package com.canoe.telegram.scenarios
 import cats.Id
 import cats.effect.IO
 import com.canoe.telegram.models.PrivateChat
-import com.canoe.telegram.models.messages.TextMessage
+import com.canoe.telegram.models.messages.{TelegramMessage, TextMessage}
 import org.scalatest.FunSuite
 import fs2.{Pure, Stream}
 
@@ -19,10 +19,10 @@ class ScenarioSpec extends FunSuite {
     TextMessage(-1, PrivateChat(chatId, None, "", None), 0, text)
 
   test("Receive >>= Expect") {
-    val scenario: Scenario[Pure, Long] =
+    val scenario: Scenario[Pure, TextMessage, Long] =
       for {
-        m <- Receive { case m: TextMessage => m.text == "one"; case _ => false }
-        _ <- Expect { case m: TextMessage  => m.text == "two"; case _ => false }
+        m <- ChatScenario.start { case m: TextMessage if m.text == "one" => m }
+        _ <- ChatScenario.next { case m: TextMessage  if m.text == "two" => m }
       } yield m.chat.id
 
     val chatId = 12
@@ -32,10 +32,10 @@ class ScenarioSpec extends FunSuite {
   }
 
   test("Scenario doesn't ignore the element which is mismatched") {
-    val scenario: Scenario[Pure, Long] =
+    val scenario: Scenario[Pure, TelegramMessage, Long] =
       for {
-        m <- Receive { case m: TextMessage => m.text == "one"; case _ => false }
-        _ <- Expect { case m: TextMessage  => m.text == "two"; case _ => false }
+        m <- ChatScenario.start { case m: TextMessage if m.text == "one" => m }
+        _ <- ChatScenario.next { case m: TextMessage  if m.text == "two" => m }
       } yield m.chat.id
 
     val input = Stream(
