--- src/test/scala/com/canoe/telegram/scenarios/ScenarioTestInstances.scala
@@ -1,8 +1,0 @@
-package com.canoe.telegram.scenarios
-
-import cats.Id
-import org.scalacheck.{Arbitrary, Gen}
-
-object ScenarioTestInstances {
-  implicit def scenarioArbitrary[A: Arbitrary]: Arbitrary[Scenario[Id, A]] = ???
-}
